[refactor] BindingAdapter 코드를 하나의 파일로 통합
