--- app/src/main/java/com/bodan/maplecalendar/presentation/views/equipment/EquipmentBindingAdapters.kt
@@ -1,9 +1,0 @@
-package com.bodan.maplecalendar.presentation.views.equipment
-
-import android.widget.TextView
-import androidx.databinding.BindingAdapter
-
-@BindingAdapter("app:grade_color")
-fun TextView.bindTextColor(colorResId: Int) {
-    setTextColor(resources.getColor(colorResId, context.theme))
-}
